OBPIH-1995 Unable to open Inventory Summary Reports

+++ grails-app/domain/org/pih/warehouse/inventory/InventoryLevel.groovy
@@ -72,7 +72,7 @@ class InventoryLevel {
         preferred(nullable: true)
 	}
 
-    def statusMessage(Integer currentQuantity) {
+    def statusMessage(Long currentQuantity) {
         return InventoryUtil.getStatusMessage(status, minQuantity, reorderQuantity, maxQuantity, currentQuantity)
     }
